comm rx

+++ application/MicroVarStore/micro_var_store.cpp
@@ -2,14 +2,14 @@
 
 #include <cstring>
 
-uint32_t MicroVarStore::get(uint8_t index) const {
+uint64_t MicroVarStore::get(uint8_t index) const {
   if (index >= MICRO_VAR_SLOT_COUNT) {
     return 0;
   }
   return vars_[index];
 }
 
-void MicroVarStore::set(uint8_t index, uint32_t value) {
+void MicroVarStore::set(uint8_t index, uint64_t value) {
   if (index >= MICRO_VAR_SLOT_COUNT) {
     return;
   }
