+++ application/MicroVarStore/micro_var_store.hpp
@@ -6,12 +6,12 @@
 
 class MicroVarStore {
  public:
-  uint32_t get(uint8_t index) const;
-  void set(uint8_t index, uint32_t value);
+  uint64_t get(uint8_t index) const;
+  void set(uint8_t index, uint64_t value);
   void clearAll();
 
  private:
-  uint32_t vars_[MICRO_VAR_SLOT_COUNT] = {};
+  uint64_t vars_[MICRO_VAR_SLOT_COUNT] = {};
 };
 
 #endif  // MICRO_VAR_STORE_HPP_
